feat(seccomp): update seccompiler to use libseccomp
libseccomp provides better quality compiler for
bpf seccomp programs than our current implementation.
The only tricky thing with this transition is the way `ioctl`
syscalls are checked with libseccomp. It always adds a check
for high bits of the request to be 0. Because of this, we need
to replace `Eq` with `MaskedEq` with mask `0x00000000FFFFFFFF`
for `ioctl` syscall.

This commit also removes dependency of firecracker and vmm
crates on the seccompiler crate.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>