fix: make aarch64_with_sve_and_pac template work with >=2 vcpus
Setting the least significant 4 bits to zero overwrites some
configuration that Firecracker sets on secondary CPUs that is needed for
them to be able to be booted. With these 4 bits set to 0, the CPUs will
never show up as online inside the guest, and KVM will just infinitely
spin inside KVM_RUN for them. Fix by using 'x' in the template, which
preserves the defaults set by Firecracker.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>