chore(unit-tests): disable debug-assertions in test builds
We need to do this because we have tests which create
Files from a raw fds or do an explicit close on fds.
This causes UB check to trigger and test to panic.
There is no separate flag to disable these checks,
so we have to disable all debug-assertions in test builds.
Tracking issue: https://github.com/rust-lang/rust/issues/123499

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>