devtool: do not call ensure_ci_artifacts in cmd_sh
This already caused a recursive call between ensure_ci_artifacts and
cmd_sh (since the former calls the later), which was only benign because
ensure_ci_artifacts exits early if the directory exists. With the next
patch, it will use cmd_sh before creating the directory, so break this
recursion once and for all by instead making the call-sites of cmd_sh
that require artifacts to just ensure them themselves.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>