feat: remove libseccomp rust dependency
Replace `libseccomp` with in house binding.
This way we don't need to add another dependency
and allows us to have access the `seccomp_export_bpf_mem`
method, not exposed in the `libseccomp` crate.
This creates another issue though: the `seccomp_export_bpf_mem`
function needs to be exposed by the libseccomp library. This
is not an issue when `seccompiler` is build in the docker environment
because we build it from source. But version provided by
linux distribution might have this function not exposed.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>