test: close stdout/stderr after timeout in utils.run_cmd
If a command times out in utils.run_cmd, we kill the subprocess and try
to get whatever partial output it wrote so far to generate an error
message. We do this using `communicate`, which waits for stdout/stderr
to close.

Some processes pass their stdout and stderr to their children. In this
case, killing the parent won't close the stdout/stderr, and so the
proc.communicate() call in run_cmd will wait indefinitely (or until
whatever is holding on the the other end of the pipe exits). Avoid this
by explicitly closing our end of the stdout/stderr pipes.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>