test: set default timeout of 100s for ssh commands
Opportunistically set a 100s timeout. We don't have anything that should
run this long, and by manually setting a timeout we avoid hitting the
pytest timeout of 300s in case something _does_ go wrong (and hitting
the pytest timeout gives significantly less debug information, so a
manual timeout is preferred).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>