test: add support for running commands with nightly toolchain
Add a `nightly` flag to the `cargo` function, which causes the requested
command to be executing using the nightly toolchain.

We have a nightly toolchain in our container due to kani. But since the
toolchain version is prescribed by kani, we need to dynamically
determine its name from `rustup toolchain list`, and cannot just use
`+nightly`.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>