fix(test): do not use absolute path to true when opening ssh connection
In #4955 the executable to check the ssh connection liveliness was
changed from `true` to `/usr/bin/true`, but that is not its path in all
rootfs, causing failures in the `test-populat-containers` suite.

Also, since the error is retried but the daemon is not cleaned
up, subsequent retries would fail for the assertion.

This change fixes both issues by using the binary name `true` and
stopping the daemon on error before the next retry.

Fixes: 3b2c2d4d6 ("test: use single SSH connection for lifetime of microvm")
Signed-off-by: Riccardo Mancini <mancio@amazon.com>