fix: stop messing with C-states in performance tests
When running performance tests, the devtool script was disabling lower
C-states, leaving only C0 (e.g. "CPU actively polling for new
instructions") enabled, with the goal to reduce latencies of switching
between C-states, and therefore increase test stability. However, we
found out that on m6i (Icelake) this was causing performance to become
erratic, with most performance metrics taking on bi-modal
characteristics.

Since back when we originally disabled sleep states, we did not see any
impact on metrics, the theory that it helps with test stability doesn't
have any experimental support, so removing the C-state configuration to
fix the volatility on m6i is not expected to have any drawbacks.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>