test: add integration test for physical counter reset
Add a test to verify the reset of the physical counter on aarch64 VMs.
To do this we check registers saved in the snapshot and verify
the counter value is less than some reasonably small number we choose.
The value is based on the observation of how much cycles it takes for
a VM to boot and be snapshotted. The idea is that this value will always
be smaller than the actual physical counter on the host.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>