devtool: Build A and B binaries if $do_build is true
If `cmd_test` is instructed to build its own binaries (e.g. being ran
outside of CI), and the environment specifies that A/B-tests should
be ran, then we not only need to build the currently checked out commit
(the "B" revision), but also an "A" binary based on the value of the
`BUILDKITE_PULL_REQUEST_BASE_BRANCH` environment variable.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>