refactor: drop `offset` field from `GuestMemoryRegionState`
In praxis, the way we wrote our snapshot files has always been just
writing all regions in-order. This mean that the offset of a region is
simply the sum of the sizes of the preceding regions. The new
`GuestMemoryMmap::create` code already computes the offsets for mapping
the memory file this way, so drop the explicit calculation at snapshot
creation time (as the calculated value isnt used by the restoration
anymore).

Do not bump the snapshot version number, because we already did so since
the last release.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>