fix(test): have codecov uploader exit with non-zero code on error
Pass the `-Z` flag (fail on error) to the codecov uploader. According to
the docs, without this flag upload failures will be silently ignored
(why isn't this the default??) [1]. We can also see this from the logs,
which contain:

```
[2025-01-28T12:10:57.973Z] ['info'] Codecov will exit with status
code
0. If you are expecting a non-zero exit code, please pass in the `-Z`
flag
```

The docs say that a long form `--fail-on-error` is available, but
actually it doesnt seem to work, so we have to use `-Z`.

[1]: https://docs.codecov.com/docs/cli-options

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>