doc: clarify cross-kernel snapshot compatibility
Simplify by just making the statement that we do not recommend it, and
it generally doesnt work except for one limited case. Simplify the table
by only listing what's working, an by having the first column directly
refer to instance types (instead of cpu architectures, which we later
map to instance types in the text).

While we're at it, also update the table of contents at the top to
include two sections it was missing.

Closes #4912
Signed-off-by: Patrick Roy <roypat@amazon.co.uk>