test: stop printing host dmesg on test failure
Because we're storing dmesg in a local variable, if an exception occurs
while killing a microvm, python will end up printing the entirety of the
host dmesg output to stderr, because it dumps local variables as part of
the backtrace. This is fine in CI because instances are shortlived, but
when running tests locally, it takes my terminal multiple minutes to
catch up with the amount of text that ends up being printed.

Fix this by not storing dmesg in a local variable anymore, and instead
inline it to its only use-site.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>