test: skip test_gitlint on PR #5048
PR #5048 is the merge of the long-lived PVH feature branch. The commits
on this branch were made long before we changes the gitlint rules to
more closely follow Linux rules when it comes to sign-offs from
co-authors (as used to not only not require them, but not allow them in
the first place, meaning the first 3 commit in this PR are only signed
by of Colin and me, but not the coauthor from a few years ago).

Explicitly skip this test for this one PR.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>