fix: Add keyboard drivers to CI guest kernels
Firecracker supports sending a CTRL+ALT+DEL event to the guest, to
trigger a reboot, and we had an integration test for this. However, the
integration test was broken (see #5050), and actually, our guest kernels
dont have the necessary drivers compiled in to actually receive the
CTRL+ALT+DEL.

Fix this by actually enabling the relavant Kconfigs. We don't re-build
the artifacts just yet, as it'll happen anyway in a few days when we
branch off the next release.

Suggested-by: Riccardo Mancini <mancio@amazon.com>
Signed-off-by: Patrick Roy <roypat@amazon.co.uk>