fix(test): install signal mask in fast_page_fault_helper
Without calling sigprocmask, the SIGUSR1 that we're using will actually
just kill the process instead of waking it up, as sigwait can only be
used to wait for _pending_ signals, but signals are only marked as
pending if they're blocked by a signal mask. If they're not blocked,
they will instead call signal handlers, which will in this case just
kill the process.

This does mean that this program never worked the intended way, which
also kinda proves it was not needed to show the different in page faults
between huge pages and normal pages (test_ept_violation_count).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>