refactor(uffd): Make example UFFD fast in absense of balloon device
Make the example UFFD handler use a HashSet to only store address ranges
that were actually removed due to expanding memory balloon. This
eliminates the HashMap that previously held entries for every single
page in guest memory, which was causing the example uffd handler to slow
to a crawl (taking about 40s to fault in 128MB of memory).

This is possible because the UFFD handler doesn't actually care about
the 4 states a page can supposedly be in, it only needs to know whether
a page needs to be served from the memory snapshot file, or zeroed out.
Since zero-ing out is the rarer of the two (only happens if a balloon
device caused some previously faulted in memory to be madvise'd away),
we only store information about pages that we got remove messages about,
and simply assume every other page needs to be served from the snapshot
file. This means that in the happy path of "no balloon device", we never
need to store any information about the memory at all.

This one simple optimization brings the fault time down to 500ms (from
40s) for the 128MB scenario, which makes the example handler usable for
performance tests for page fault latencies.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>