refactor(test): Store binary_dir inside microvm
Just store the binary dir inside the microvm class instead of separate
paths to firecracker and jailer binaries. This centralizes the existance
assertions, but means we cant theoretically use jailer and firecracker
binaries from different locations anymore (however, nothing in the test
framework was doing this, as MicroVMFactory already did not support this
and hardcoded the binary names relative to the --binary-dir option).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>