doc(vsock): clarify vsock support on snapshot/restore
The current documentation is not clear, mentioning that the device could
break if active at time of snapshot. It also mentions that this is fixed
by resetting the device at snapshot time, closing all open connections.

This patch clarifies that vsock connections are explicitly closed by a
transport reset, in order to avoid the aforementioned issue. This
doesn't impact listening sockets as they are able to accept new
connections after restore.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>