fix: do not try to use put_dimensions to merge dimension sets
That's apparently not how that function works. Instead, it just
associates a completely separate dimension set with the
metric-to-be-emitted, and then emits it twice (with different dimensions
sets). Needless to say, that's not what we want, so go back to centrally
configuring all metrics with a single call to set_dimensions.

Fixes: 31ee8512e50e
Signed-off-by: Patrick Roy <roypat@amazon.co.uk>