Revert "fix: test_sec_audit: ignore cargo audit output"
This reverts commit d761b013db1caeec046b187927133b7fdac1375d. The output
to stdout was fixed in cargo audit 0.21.2, so if we rebuild the docker
container the grep is no longer necessary.

In fact, the grep has broken this test in our nightly pipeline because
it overwrites the return code of cargo audit itself, meaning the non-PR
version of this test (which is supposed to fail if there exist any cargo
audit warnings) was never failing.

(cherry picked from commit 1d98a21e087aa4f188f2d6e665b54c6962569199)
Signed-off-by: Patrick Roy <roypat@amazon.co.uk>