fix(vmm): Calc right-shift bits to address socket ID
The number of bits that the x2APIC ID must be shifted to the right to
address instances of the socket domain, which is next higher-scoped
domain to the core domain, was hardcoded as 7. That means up to 128
vcpus support. Currently the max vcpu count is hardcoded as 32, so it
is enough but we might want to increase it in the future. To avoid
unexpected issues when increasing the max vcpu count, calculate the
value based on the max vCPU count.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>