doc: update quickstart guide to download latest artifacts
We need to manually change what versions our quick start guide points to
at every release. To avoid it getting behind, this change adds automatic
detection of the CI artifacts related to the lastest version (which was
already automatically detected).

This means that CI artifacts will not point to the latest available but
they will be the same we use to test the latest release, to ensure
compatibility and that the guide always works.

Signed-off-by: Riccardo Mancini <mancio@amazon.com>