chore: update dependencies to appease `cargo audit`
`cargo audit` warns us about a security issue in `hashbrown` crate
(RUSTSEC-2024-0402). We use this as a development dependency, so it
does not affect Firecracker in production. As a result, we don't need to
create a patch release.

`cargo audit` also emits a warning for `kvm-ioctls` about
RUSTSEC-2024-0428 which has to do with how newer Rust compilers handle
downcasts mutable references to immutable pointers. Starting from 1.82.0
Rust compilers elide subsequent reads of this struct's fields post
downcast. However, here we are using Rust 1.79.0, so this is not a
problem. We update the dependency to appease `cargo audit`.

Signed-off-by: Babis Chalios <bchalios@amazon.es>