ci: Allow build script to run without install
Update the script to allow it to run the build without installing.

Allow us to add patches in different formats, we now allow .lore
files which are a link to the patch series, these will be automatically
downloaded and applied. We also accept .mbox files which again will be
patched onto our kernel.

Signed-off-by: Jack Thomson <jackabt@amazon.com>