fix(test): make test_cargo_audit not ignore return code outside A/B
The pipe through grep was causing any return code of cargo audit to be
overwritten with that of grep (which is always 0). Fix this by using `-o
pipefail`, which changes the return code of a bash pipe to instead be
that of the rightmost failing command (or 0 if none fail). Note that
just -o pipefail will not abort the pipe on a non-zero code (so the grep
will still run even if cargo audit fails).

Suggested-by: Riccardo Mancini <mancio@amazon.com>
Signed-off-by: Patrick Roy <roypat@amazon.co.uk>