doc: clarify kernel configuration usage in our repo
We keep in the repo the kernel configurations from which we build the
kernel images we use in our CI pipelines. However, we don't mention what
kernel sources we use to build these kernel images.

Clarify in the documentation that these configurations are meant to be
used with kernels provided by Amazon Linux and that we don't guarantee
that they can be used to build kernels from upstream Linux.

Signed-off-by: Babis Chalios <bchalios@amazon.es>