add function to VmResources for allocating I/O memory
If swiotlb is requested through the API, this function will allocate a
single contiguous region intended for swiotlb use. It needs to be
continguous because we must be able to describe it using a single memory
range in FDT, as devices cannot be assigned to reserved memory
consisting of multiple regions (e.g. FDT assignment is to memory
regions, not to #reserved-memory nodes).

While we're at it, always use anon memory for the "normal" part of guest
memory if swiotlb is enabled, as if swiotlb is enabled we know that
vhost devices will never need to access that part of memory.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>