refactor: use #[from] in VmmError
When a variant wraps another error type, and this error type does not
appear in any other variant, use thiserrors #[from] directive to
implement a `From` impl. This allows eliminating almost all remaining
`map_err`s from `create_vmm_and_vcpus`, as the `?` operator will
automatically call .from() once (although it cannot do it for multiple
layers of wrapping, which is why it was important to change the function
to return VmmError).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>