ci: Add secret hiding kernel to defaults buildkite
Adding the secret hiding kernel as a default for the buildkite pipeline,
this will mean that PR's made against the branch will now be run with
the new secret hiding enabled amis.

Some tests have been marked to skip as they are kernel dependent so
while we are compiling our kernel in CI these could change again.

Signed-off-by: Jack Thomson <jackabt@amazon.com>