have virtio devices offer VIRTIO_F_ACCESS_PLATFORM if swiotlb
Have all virtio devices off VIRTIO_F_PLATFORM_ACCESS, which communicates
to the guest that DMA APIs should be used for allocating vrings and
virtio buffers. Since we have created a dedicated swiotlb region in the
geust and assigned all virtio devices to this, it means that all such
allocations will go through the swiotlb region.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>