Snapshot restore for swiotlb regions
Support restoring VMs with swiotlb regions. For this, untangle the uffd
handshake from the actual restoration of memory regions, as we first
need to restore and register to the Vm _all_ memory regions, before we
can then send a single handshake containing both normal and swiotlb
regions to the Uffd handler.

While we're here, significantly simplify the jungle of error types.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>