refactor: hoist post-snapshot vring re-dirtying
Do this in create_snapshot() instead of snapshot_memory_to_file(), as it
was the only operation inside snapshot_memory_to_file() that required
access to a struct Vmm, opening the gates for move the rest of
snapshot_memory_to_file() into struct Vm.

Also change the error type as to not drag in VmmError into the vm
module, where it really doesn't belong.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>