chore: allow clippy::needless_update
This lint forbids using `..Default::default()` in struct initializers
after all fields have already been initialized, but this is a useful
pattern if you know you want to add more fields to a struct in a future
PR without needing to touch a ton of initializers in unittests again
(_heavy foreshadowing_). So silence the paperclip.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>