fix: dont fail silently on awscli errors during artifact setup
If something goes wrong during aws s3 cp, exit early with an error
message instead of quietly trying to soldier on. A common(-ish) cause of
something going wrong here is having installed the aws cli for the wrong
CPU architecture (e.g. x86 AWSCLi on ARM64), as the awscli installer
apparently doesnt not check the architecture, and its easy to copy the
wrong command from the website.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>